// src/app/shared/model/ref-office-obj.model.ts
export class RefOfficeObj {
    RefOfficeId: number;
    OfficeCode: string;
    OfficeName: string;
    OfficeShortName: string;
    MrKonvenSyariahCode: string;
    OfficeAddr: string;
    Zipcode: string;
    AreaCode1: string;
    AreaCode2: string;
    AreaCode3: string;
    AreaCode4: string;
    PhnArea1: string;
    Phn1: string;
    PhnExt1: string;
    PhnArea2: string;
    Phn2: string;
    PhnExt2: string;
    PhnArea3: string;
    Phn3: string;
    PhnExt3: string;
    FaxArea: string;
    Fax: string;
    CntctPersonName: string;
    CntctPersonJobTitle: string;
    CntctPersonEmail1: string;
    CntctPersonEmail2: string;
    CntctPersonMobilePhnNo1: string;
    CntctPersonMobilePhnNo2: string;
    MrOfficeClassCode: string;
    RefOfficeAreaId: number;
    ParentOfficeCode: string;
    IsActive: boolean;
    ParentId: number;
    IsOfficeClose: boolean;
    OfficeOpeningDt: Date;
    IsAllowAppCreated: boolean;
    HolidaySchmHId: number;
    WorkingHourSchmHId: number;
    IsVirtualOffice: boolean;
    MrOfficeTypeCode: string;
    MrCenterGrpTypeCode: string;
    RowVersion: string;
    constructor() { this.RefOfficeId = 0; this.RowVersion = '' }
}